+++ es/data_entry.go
@@ -1,0 +1,7 @@
+package es
+
+// DataEntry represents data entry
+type DataEntry struct {
+	Name  string `json:"name"`
+	Value string `json:"value"`
+}
