+++ es/price.go
@@ -1,0 +1,7 @@
+package es
+
+// Price represents Price struct from XDR
+type Price struct {
+	N int `json:"n"`
+	D int `json:"d"`
+}
