+++ es/signer.go
@@ -1,0 +1,7 @@
+package es
+
+// Signer represents signer as export
+type Signer struct {
+	Key    string `json:"key"`
+	Weight int    `json:"weight"`
+}
